Development (#205)
* Merge (#190)

* Global js beautify (#158)

* running all the js files through js-beautify using the paramaters given on the website

* updating bracket to expanded, added npm script

* cannot decide on brace_style

* beautifying

* updating CONTRIBUTING docs on javascript formatting

* updating package lock file

* rebeauitifying

* add note and link to product ideas document (#175)

* Beautify public js (#171)

* beautify editcube.js

* beautify files in public/js

* beautify public/js automatically (#170)

* install beautify during npm install

* add public/js to readme instructions on beautify

* include public/js in default beautify command

* revert unnecessary changes

* deduplicate card addition (#166)

* implement util.addCardToCube

* call util.addCardToCube where appropriate

* call util.addCardToCube in special case

* re-add missing C (#194)

* Fixed issue with edit token

* fix compare columns on mobile (#199)

* changed wording edit selected (#201)

This is an attempted resolution to issue#160. This update changes the "Edit Selected" button to say "Mass Edit" while
the user is not in list view, and changes it to say "Edit Selected"
while they are in list view. In addition, pressing the button while not
in list view will switch the user to list view instead of showing a
dialog.

I am not sure about my implementation for changing the text of the
button. An alternative might be to put some javascript into the pug
file, to be evaluated on the page as the user switches views. This might
be preferable as it keeps all display related code in the pug file
(cube_list.pug) rather than in the javascript.

I have also changed the name for the DOM element "editSelected" to
"massEdit" to reflect it's new default value.

* Fix dev (#203)

* Global js beautify (#158)

* running all the js files through js-beautify using the paramaters given on the website

* updating bracket to expanded, added npm script

* cannot decide on brace_style

* beautifying

* updating CONTRIBUTING docs on javascript formatting

* updating package lock file

* rebeauitifying

* add note and link to product ideas document (#175)

* Beautify public js (#171)

* beautify editcube.js

* beautify files in public/js

* beautify public/js automatically (#170)

* install beautify during npm install

* add public/js to readme instructions on beautify

* include public/js in default beautify command

* revert unnecessary changes

* deduplicate card addition (#166)

* implement util.addCardToCube

* call util.addCardToCube where appropriate

* call util.addCardToCube in special case

* Merge (#190) (#191)

* Global js beautify (#158)

* running all the js files through js-beautify using the paramaters given on the website

* updating bracket to expanded, added npm script

* cannot decide on brace_style

* beautifying

* updating CONTRIBUTING docs on javascript formatting

* updating package lock file

* rebeauitifying

* add note and link to product ideas document (#175)

* Beautify public js (#171)

* beautify editcube.js

* beautify files in public/js

* beautify public/js automatically (#170)

* install beautify during npm install

* add public/js to readme instructions on beautify

* include public/js in default beautify command

* revert unnecessary changes

* deduplicate card addition (#166)

* implement util.addCardToCube

* call util.addCardToCube where appropriate

* call util.addCardToCube in special case

* Fix spelling of cookie and lengthen expiration. (#197)

app.js misspelled cookie, and so the secure attribute was not actually being set. Additionally, the maximum age was not being set, and so logouts occur frequently. On the current site, this often leads to putting in a bunch of changes and then having them disappear because the "save" button just leads to a re-login.

* Development (#202)

* Merge (#190)

* Global js beautify (#158)

* running all the js files through js-beautify using the paramaters given on the website

* updating bracket to expanded, added npm script

* cannot decide on brace_style

* beautifying

* updating CONTRIBUTING docs on javascript formatting

* updating package lock file

* rebeauitifying

* add note and link to product ideas document (#175)

* Beautify public js (#171)

* beautify editcube.js

* beautify files in public/js

* beautify public/js automatically (#170)

* install beautify during npm install

* add public/js to readme instructions on beautify

* include public/js in default beautify command

* revert unnecessary changes

* deduplicate card addition (#166)

* implement util.addCardToCube

* call util.addCardToCube where appropriate

* call util.addCardToCube in special case

* re-add missing C (#194)

* Fixed issue with edit token

* fix compare columns on mobile (#199)

* changed wording edit selected (#201)

This is an attempted resolution to issue#160. This update changes the "Edit Selected" button to say "Mass Edit" while
the user is not in list view, and changes it to say "Edit Selected"
while they are in list view. In addition, pressing the button while not
in list view will switch the user to list view instead of showing a
dialog.

I am not sure about my implementation for changing the text of the
button. An alternative might be to put some javascript into the pug
file, to be evaluated on the page as the user switches views. This might
be preferable as it keeps all display related code in the pug file
(cube_list.pug) rather than in the javascript.

I have also changed the name for the DOM element "editSelected" to
"massEdit" to reflect it's new default value.

* Fix login issue