Remove Edit Token
Edit Token was unnecessary for the app to work as intended, and just
added complexity to any possible new routes. Now validity is checked on
the server side based on whether the de-serialized user object generated by
passport matches the cube owner, and is checked on the client side by a
hidden input field with a generic value. That input field is only
created if a user exists as a part of the request. This has no changed
front-end functionality, as currently the front-end only checked whether
the edit token existed, then simply passed it back to the server.

No additional token needs to be passed back and forth, now the session
cookie handles authentication, like it should.