Feature 3 fix2 (#327)
* tokenize bugfix and sort bugfix

Fixed two bugs. In editcube, the tokenize fucntion wouldn't properly
return the value of the yet to be filtered text, instead truncating it
at the next occurence of token.arg. This was particularly problematic
for short/common args like "r" for the color red.

In SortFilter, I forgot to add an argument to a function call, and it
has been added now.

* fix other instance of missing args

* remove logs that were slowing search

A ton of console.logs being called for debugging purposes were vastly
slowing down search. I've left in some that will help me with debugging
while on dev, but disable a vast majority. This should improve
performance quite a bit!