add explanation for parsing logic
This is a very informal attempt at documenting the parsing logic, in the hopes that other people will be able to work on and contribute to it. I hope that people will update this document when they do so, since it contains many suggestions that could be implemented. Currently, they are simply improvements to how the code could be cleaner, not actually bugs, so I'd rather not submit them as issues, though I'm open to discussion on that point.