Fix bottom of autocard popup lingering
This issue was caused by the image unloading handler being run
erroneously after the autocard_hide handler attempts to remove its
events. This is difficult to spot in most cases but becomes easily
reproducible if the tester disables caching and sets throttling to
something very low (on Firefox/Windows 10, "Regular 2G" did the trick...
set this after the cube list loads or you could be waiting a while).
On fast network connections, this would rarely be an issue, as the image
would load properly faster than the user could move their mouse off the
image, and browser caching prevents them from trying again on the same
card.

The issue was that innerHtml on the image container was being cleared
before the event handlers were canceled. Clearing this property removes
the child nodes from the document, meaning the queries in the lines
afterward do not find them and thus do not remove the event handler
(jquery, helpfully, does not indicate to the user that nothing was
found) and it fires regardless, showing the (now-empty) popup if the
tags variable was captured with something in it.

Clearing the events properly fixes the bug, and it can no longer be
reproduced even with throttling/nocache.