Cache a promise for generating pack images
This means that when N requests for the same image come in quick
succession (for example, when a pack image is displayed on Discord),
the server doesn't need to render the image N times. Instead, it
renders it once, then sends the resulting data out N times,
substantially decreasing the server load and increasing
responsiveness.