Update to fleet config for elastic-agent.
The non-fleet elastic agent doesn't work on autopilot since it tries to mount a hostPath volume for writing.
See #elastic/cloud-on-k8s#4699 for details.
Unfortunately fleet support is only available on master of cloud-on-k8s and not really documented.
Will wait to deploy this until we come up with a decision on if we want to use an experimental build.